refactor: move REASONING_FEATURE_LABEL out of shared into reasoning-toggle
The label is domain vocabulary used by only 2 slices, which violates
shared/CLAUDE.md rules #1 (no domain vocab in shared) and #4 (1-2 slice
utils live in the slice, promote after the 3rd use). Move it into its
owner slice features/reasoning-toggle (barrel-exported); analysis-nudge
imports it cross-feature. Single source of truth preserved (no drift).
Documents the analysis-nudge → reasoning-toggle exception in
src/features/CLAUDE.md.
