+++ src/features/reasoning-toggle/index.ts
@@ -1,3 +1,4 @@
 // reasoning-toggle feature barrel — member "깊은 생각" toggle 공개 API.
 export { ReasoningToggle } from './ui/ReasoningToggle';
 export { useReasoningToggle } from './hooks/useReasoningToggle';
+export { REASONING_FEATURE_LABEL } from './model/reasoningFeature';
